Varint fixes for ints, uints, sints, and enums (#38)
This PR reworks varint decoding and encoding. The previous approach was
flawed because (1) it did not cast to expected ranges when decoding, (2)
decoding and encoding lost more precision than necessary in some cases,
and (3) decoding made incorrect assumptions about the number of bytes
read and whether a value was negative.

The new approach implements (1) separate decoders for each of the
numeric datatypes that use varint, (2) an improved writer that preserves
more precision with bigger numbers, and (3) a separate writer for
zigzag-encoded values (sint32 and sint64) to offer better precision
guarantees. This is also why the `limitInt32` and `*zigzag` functions
have been removed.

The new approach passes 38 conformance tests that previously failed!

There are 4 regressions in the tests, but these tests should never have
passed in the first place - it was only due to a bug that they did. Each
test input was varint(max_uint64) and we were expected to round-trip it,
which we can't (without a special 64-bit int representation). The tests
passed because the varint was read as -1 which happens to reencode as
varint(max_uint64), the expected output.